Tweaks to styling and fetching
